Added: Generic presenters don't work do to the limitation that inject(BaseViewFragment<MainFragmentPresenter> bvf) hard type is required, no <? extends BasePresenter> is allowed thus you will have to list all of the injections required instead of inheriting
See: https://gist.github.com/konmik/6ac725fa7134402539c4

Will need to create an @Module with @Provides if scoping becomes a requirement for the presenters, thus currently each individual Presenter class requires the @Inject constructor, but would be replaced with the @Provides returning the scoped instance(unless you would like to annotate the class), however I like the decoupling of the dependent class from the annotations, to which the module would be placed under the fragment/whatever scope.

Via the current implementation, you can provide application scoped thread handlers as a constructor dependency for each presenter such that a rotation will create the new screen specific presenter, along with populating it with the thread handler allowing the presenter to handle the long running call(which exist across configuration changes). This would require an interface on the presenters, with the subscription code to the long running task, and a queue for the task result if the task completes when the device rotates or the presenter is destroyed(null) due to some other reason, allowing the data to be retrieved when the presenter is ready.

Another consideration is persisting the tasks data/or the presenters result(based on where you want to save) such that the presenter can receive data, get destroyed and receive the data from a cache/long term persistence.